+++ server/src/main/java/org/allaymc/server/block/component/BlockFireBaseComponentImpl.java
@@ -281,7 +281,7 @@ public class BlockFireBaseComponentImpl extends BlockBaseComponentImpl {
             }
 
             if (target.getBlockType() == BlockTypes.TNT) {
-                BlockTypes.TNT.getBlockBehavior().prime(target);
+                BlockTypes.TNT.getBlockBehavior().prime(target.getPosition());
             }
         }
     }